imara_diff::sink

Struct Counter

pub struct Counter<T> {
    pub removals: u32,
    pub insertions: u32,
    pub wrapped: T,
}

A Sink which wraps a different sink and counts the number of removed and inserted tokens.

Fields
removals: u32

Total number of recorded inserted tokens. Computed by summing the lengths of the after subsequences pass to process_change.

insertions: u32

Total number of recorded inserted tokens. Computed by summing the lengths of the after subsequences pass to process_change.

wrapped: T

The Sink for which the counter records tokens. All calls to process_change are forwarded to the sink by the counter. After finish is called, this field contains the output returned by the finish method of the wrapped Sink.

Implementations

impl<S: Sink> Counter<S>

pub fn new(sink: S) -> Self

impl<T> Counter<T>

pub fn total(&self) -> usize