imara_diff
Module sources

Structs

ByteLines

A TokenSource that returns the lines of a byte slice as tokens. See byte_lines for details.

Lines

A TokenSource that returns the lines of a str as tokens. See lines for details.

Functions

byte_lines

Returns a TokenSource that uses the lines in data as Tokens. The newline separator (\r\n or \n) is included in the emitted tokens. This means that changing the newline separator from \r\n to \n (or omitting it fully on the last line) is detected when computing a Diff.

lines

Returns a TokenSource that uses the lines in data as Tokens. The newline separator (\r\n or \n) is included in the emitted tokens. This means that changing the newline separator from \r\n to \n (or omitting it fully on the last line) is detected by Diff.