imara_diff::sources
Function lines

pub fn lines(data: &str) -> Lines<'_> 

Returns a TokenSource that uses the lines in data as Tokens. The newline separator (\r\n or \n) is included in the emitted tokens. This means that changing the newline separator from \r\n to \n (or omitting it fully on the last line) is detected by Diff.