imara_diff
Struct Interner

pub struct Interner<T> { /* private fields */ }

An interner that allows for fast access of tokens produced by a TokenSource.

Implementations

impl<T> Interner<T>

pub fn new_for_token_source<S: TokenSource<Token = T>>( before: &S, after: &S, ) -> Self

Create an Interner with an initial capacity calculated by summing the results of calling estimate_tokens methods of before and after.

pub fn new(capacity: usize) -> Interner<T>

Create an Interner with initial capacity capacity.

pub fn clear(&mut self)

Remove all interned tokens.

pub fn num_tokens(&self) -> u32

Returns to total number of distinct tokens currently interned.

impl<T: Hash + Eq> Interner<T>

pub fn reserve_for_token_source<S: TokenSource<Token = T>>( &mut self, before: &S, after: &S, )

Create an Interner with an intial capacity calculated by calling estimate_tokens methods of before and after

pub fn reserve(&mut self, capacity: usize)

pub fn intern(&mut self, token: T) -> Token

Intern token and return a the interned integer.

pub fn erase_tokens_after(&mut self, first_erased_token: Token)

Erases first_erased_token and any tokens interned afterward from the interner.